refactor(actions): split role-submissions.ts and enrichment.ts into responsibility modules
Behavior-preserving structural refactor. The two largest action files
(1254 + 917 lines) become directories with responsibility-grouped sub-modules
and a re-export barrel that preserves the existing `@/actions/role-submissions`
and `@/actions/enrichment` import paths.

role-submissions/
  shared.ts     types + Zod schemas + STALE_STATUSES / CUSTOMER_ALLOWED
  mutations.ts  submitCandidatesToCustomer / updateSubmissionStatus / removeSubmission
  sharing.ts    generateShareToken / revokeShareToken
  public.ts     updateSubmissionStatusByToken / getSubmissionByToken (no-auth)
  reads.ts      getSubmissionsForRole / getRecentSubmissionsForTenant / getAllSubmissionsForTenant
  index.ts      re-export barrel

enrichment/
  http.ts        timedFetch / readCappedText (internal helper)
  search.ts      braveSearch / classifySource / toProfileSource (internal helper)
  platforms.ts   fetchGitHubProfile / fetchStackOverflowProfile / fetchPersonalSiteSummary
  matcher.ts     buildCvSignalsForMatcher / deterministicVerify (internal helper)
  orchestrate.ts enrichCandidate + EnrichmentResult
  verdicts.ts    confirmProfile / dismissProfile
  links.ts       updateCandidateLinks
  index.ts       re-export barrel

Public API is unchanged: every name previously exported from the two source
files is re-exported from the new index barrel under the same import path.
Function bodies are byte-identical to the originals (verified by diff against
origin/core-mvp-foundation). No signature changes, no audit-log changes, no
new types extracted unnecessarily.

May conflict at merge time with PR #178 (audit middleware migration) — that's
expected and acceptable per the splitting plan.

Verification:
- `npx tsc --noEmit` → clean before and after the split
- `NODE_ENV=test npm test` → identical pass/fail set before vs after
  (9 failed files / 23 tests are pre-existing on core-mvp-foundation HEAD
  and unrelated to role-submissions or enrichment)
- Grep snapshot of consumer imports unchanged (sorted diff is empty):
  - `@/actions/role-submissions`: 12 consumer sites preserved
  - `@/actions/enrichment`: 4 consumer sites preserved
